project_feedback_ckasahara
This overall looks like a solid project! Your graphs look good and they all seem necessary for the topic you are looking at. You don't have to, but for the number of sightings, I think a bubble map would look really good to display the number of sightings per area! It would really translate how much is seen in each state. The number of UFOs per state bar plot is a little tight, so if you are not able to stretch it out due to the high number of recordings, I would either group them or use another type of graph format to translate this idea.
Great job!